--- a/modules/sd_models.py
+++ b/modules/sd_models.py
@@ -236,11 +236,15 @@ def read_metadata_from_safetensors(filename):
 
 def read_state_dict(checkpoint_file, print_global_state=False, map_location=None):
     _, extension = os.path.splitext(checkpoint_file)
-    if extension.lower() == ".safetensors":
-        device = map_location or shared.weight_load_location or devices.get_optimal_device_name()
-        pl_sd = safetensors.torch.load_file(checkpoint_file, device=device)
-    else:
-        pl_sd = torch.load(checkpoint_file, map_location=map_location or shared.weight_load_location)
+    with open(checkpoint_file, 'rb') as f:
+        if extension.lower() == ".safetensors":
+            buffer = f.read()
+            pl_sd = safetensors.torch.load(buffer)
+        elif extension.lower() == ".ckpt":
+            buffer = io.BytesIO(f.read())
+            pl_sd = torch.load(buffer, map_location)
+        else:
+            raise Exception(f"Unknown model type: {extension}")
 
     if print_global_state and "global_step" in pl_sd:
         print(f"Global Step: {pl_sd['global_step']}")