--- a/modules/sd_vae.py
+++ b/modules/sd_vae.py
@@ -120,7 +120,7 @@ def resolve_vae(checkpoint_file):
 
 
 def load_vae_dict(filename, map_location):
-    vae_ckpt = sd_models.read_state_dict(filename, map_location=map_location)
+    vae_ckpt = sd_models.read_state_dict(filename)
     vae_dict_1 = {k: v for k, v in vae_ckpt.items() if k[0:4] != "loss" and k not in vae_ignore_keys}
     return vae_dict_1