this pr changes weights loading from implict file handling by `torch` or `safetensors` to explicit binary stream  

now my local storage where models are stored is getting fully saturated during model read (>600mb/sec while previously it was fluctuating between 180-500mb/sec)

pr also removes obsolete variables, `map_location` and `weight_load_location`

keeping in draft pending some testing for possible side-effects.

## results

- **ckpt**: ~30% (**1.5x**) faster using loading using `torch.load` from io stream is than from file
- **safetensors** ~50% (**2x**) faster loading using `safetensors.torch.load` from file stream than `safetensors.torch.load_file`  

*note*: those are my worst-case examples, i've seen even bigger speedups in some tests

and reported in console - not bad for a 4GB model:

> Model loaded in 7.6s (load weights: 5.3s, create model: 0.4s, apply weights: 0.6s, load vae: 0.5s, device move: 0.8s).

## notes

- `load_file` attempts to memory map the file, but does so quite inefficiently  
  (see my notes at <https://github.com/huggingface/safetensors/issues/200#issuecomment-1478326999> for details)
- `safetensors.torch.load` does not have `device` param (unlike `load_file`) and always loads to cpu anyhow and model gets moved as needed later

## follow-ups

1. this pr can be extended to wrap `open()` function with a progress bar:

simply replace:  
> with open(checkpoint_file, 'rb') as f:

with  
> import rich  
> with rich.progress.open(checkpoint_file, 'rb') as f:

and the console output becomes:

    Loading weights from /home/vlado/dev/automatic/models/Stable-diffusion/perfect-world-v20.safetensors
    Reading... ━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━ 4.3/4.3 GB 0:00:00

2. what is the impact when loading model from non-local storage? e.g. mapped s3 bucket