@CapsAdmin 2023-03-21
This fails for me. Apart from forgetting to import io, it fails to load `models/Stable-diffusion/1.X/vae-ft-mse-840000-ema-pruned.vae.ckpt`

with the error
```
/home/caps/projects/stable-diffusion-webui/modules/safe.py:30: UserWarning: TypedStorage is deprecated. It will be removed in the future and UntypedStorage will be the only storage class. This should only matter to you if you are using storages directly.  To access UntypedStorage directly, use tensor.untyped_storage() instead of tensor.storage()
  return TypedStorage()
loading stable diffusion model: UnpicklingError
Traceback (most recent call last):
  File "/home/caps/projects/stable-diffusion-webui/webui.py", line 136, in initialize
    modules.sd_models.load_model()
  File "/home/caps/projects/stable-diffusion-webui/modules/sd_models.py", line 441, in load_model
    load_model_weights(sd_model, checkpoint_info, state_dict, timer)
  File "/home/caps/projects/stable-diffusion-webui/modules/sd_models.py", line 334, in load_model_weights
    sd_vae.load_vae(model, vae_file, vae_source)
  File "/home/caps/projects/stable-diffusion-webui/modules/sd_vae.py", line 145, in load_vae
    vae_dict_1 = load_vae_dict(vae_file, map_location=shared.weight_load_location)
  File "/home/caps/projects/stable-diffusion-webui/modules/sd_vae.py", line 123, in load_vae_dict
    vae_ckpt = sd_models.read_state_dict(filename, map_location=map_location)
  File "/home/caps/projects/stable-diffusion-webui/modules/sd_models.py", line 246, in read_state_dict
    pl_sd = torch.load(buffer, map_location=map_location or shared.weight_load_location)
  File "/home/caps/projects/stable-diffusion-webui/modules/safe.py", line 106, in load
    return load_with_extra(filename, extra_handler=global_extra_handler, *args, **kwargs)
  File "/home/caps/projects/stable-diffusion-webui/modules/safe.py", line 151, in load_with_extra
    return unsafe_torch_load(filename, *args, **kwargs)
  File "/home/caps/projects/stable-diffusion-webui/venv/lib/python3.10/site-packages/torch/serialization.py", line 817, in load
    return _legacy_load(opened_file, map_location, pickle_module, **pickle_load_args)
  File "/home/caps/projects/stable-diffusion-webui/venv/lib/python3.10/site-packages/torch/serialization.py", line 1035, in _legacy_load
    magic_number = pickle_module.load(f, **pickle_load_args)
_pickle.UnpicklingError: A load persistent id instruction was encountered,
but no persistent_load function was specified.
```