@fractal-fumbler 2023-03-21
double checked: getting error `NameError: name 'print_global_state' is not define`

<details>

```python
/tmp/stable-diffusion-webui/auto/lib/python3.10/site-packages/pytorch_lightning/utilities/distributed.py:258: LightningD
eprecationWarning: `pytorch_lightning.utilities.distributed.rank_zero_only` has been deprecated in v1.8.1 and will be re
moved in v2.0.0. You can import it from `pytorch_lightning.utilities` instead.                                          
  rank_zero_deprecation(                                                                                                
Loading weights [038d767451] from /home/server/stable-diffusion/models/checkpoints/test/test_merge_SD-var4_05.safetensor
s                                                                                                                       
loading stable diffusion model: NameError                                                                               
Traceback (most recent call last):                                                                                      
  File "/tmp/stable-diffusion-webui/webui.py", line 136, in initialize                                                  
    modules.sd_models.load_model()                                                                                      
  File "/tmp/stable-diffusion-webui/modules/sd_models.py", line 412, in load_model                                      
    state_dict = get_checkpoint_state_dict(checkpoint_info, timer)                                                      
  File "/tmp/stable-diffusion-webui/modules/sd_models.py", line 267, in get_checkpoint_state_dict                       
    res = read_state_dict(checkpoint_info.filename)                                                                     
  File "/tmp/stable-diffusion-webui/modules/sd_models.py", line 250, in read_state_dict                                 
    if print_global_state and "global_step" in pl_sd:                                                                   
NameError: name 'print_global_state' is not define
```

</details>